Allow webrtc connection to exit early
I know this isn't a fully fledged PR, I'm just seeing what you think of this

I've been running into the issue of iceGatheringState being stuck on 'gathering' even after receiving viable candidates, then hitting the timeout (10 seconds) which makes calls to /connections super slow. I'm not really sure why this is the case and based on your comment `// strangely something it takes a long time` it's the same for you. 

This allows the user to specify minimum srflx candidates that will stop iceGathering early.

The other option here would be to expose the timeout (TIME_TO_HOST_CANDIDATES) as a server option as 10 seconds is too long to wait anyway.