Allow CORS to be configured to allow Authorization
Geckos version 1.6.0 introduces support for the HTTP authorization
header on channel initialization. However, when connecting to a
cross-origin geckos server, an authorized request will get denied
because `authorization:` wasn't an allowed CORS header. This adds an
option to the Geckos server CORS configuration object that allows the
user to allow the `authorization:` header in a cross-origin context.